ai/requirements.txt
websocket-client==1.9.0
pillow>=10.0.0
termcolor>=2.3.0
python-dotenv>=1.0.0
flask
flask-cors
